@@ -3,38 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.2.0",
+    "@ant-design/icons": "^5.1.4",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
+    "axios": "0.21.2",
     "babel-plugin-import": "^1.13.3",
-    "craco-less": "^1.18.0",
     "customize-cra": "^1.0.0",
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.9.1",
     "echarts": "^5.2.0",
     "echarts-for-react": "^3.0.1",
+    "html-to-draftjs": "^1.5.0",
     "less": "^4.1.1",
     "less-loader": "^7.3.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.1.8",
-    "react-dom": "^17.0.2",
-    "react-draft-wysiwyg": "^1.14.7",
-    "react-redux": "^7.2.4",
+    "react-dom": "^18.2.0",
+    "react-draft-wysiwyg": "^1.15.0",
+    "react-redux": "^8.0.7",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.1",
+    "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
+    "redux-thunk": "^2.4.2",
     "sass-loader": "^10.1.1",
     "store": "^2.0.12",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^1.1.2",
+    "lottie-react": "^2.4.0"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "lint": "eslint --fix --ext .js,.jsx ."
   },
   "eslintConfig": {
@@ -56,10 +55,16 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-decorators": "^7.14.5",
+    "@vitejs/plugin-react": "^3.1.0",
+    "@vitejs/plugin-react-swc": "^3.3.0",
     "antd": "^4.16.8",
+    "autoprefixer": "^10.4.14",
     "eslint": "^7.32.0",
     "eslint-plugin-react": "^7.24.0",
     "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2"
+    "prettier": "^2.3.2",
+    "vite": "^4.3.3",
+    "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-svgr": "^2.4.0"
   }
 }
