@@ -1,101 +1,66 @@
-import React, { Component } from 'react';
-import { withRouter } from 'react-router-dom';
+import React, {useState, useEffect} from 'react';
+import {useHistory} from 'react-router-dom';
 import './index.less';
 import storageUtils from '../../utils/storageUtils';
 import memoryUtils from '../../utils/memoryUtils';
-import { Modal } from 'antd';
-import menuList from '../../config/menu';
+import {Modal} from 'antd';
 import LinkButton from '../link-button';
-import { connect } from 'react-redux';
+import {connect} from 'react-redux';
 
-class HeaderCom extends Component {
-    logout = () => {
-        /* must clear memoryutils and storageutils before redirect
-      or it will redirect to admin page again.
-    */
+function HeaderCom({headTitle}) {
+    const [logoutIsVisable, setLogoutIsVisable] = useState(false);
+    const [currentTime, setCurrentTime] = useState(new Date().toString());
+    const history = useHistory();
+
+    useEffect(() => {
+        const timer = setInterval(() => setCurrentTime(new Date().toString()), 1000);
+        return () => clearInterval(timer);
+    }, []);
+
+    const logout = () => {
         storageUtils.removeUser();
         memoryUtils.user = {};
-        this.props.history.push('/login');
-        // event.preventDefault()
+        history.push('/login');
     };
 
-    state = {
-        logoutIsVisable: false,
-        currentTime: new Date().toString(),
+    const handleCancel = () => {
+        setLogoutIsVisable(false);
     };
 
-    interPointer = null;
-
-    componentDidMount() {
-        this.interPointer = setInterval(() => {
-            this.setState({ currentTime: new Date().toString() });
-        }, 1000);
-    }
-
-    componentWillUnmount() {
-        if (this.interPointer) clearInterval(this.interPointer);
-    }
-
-    handleCancel = () => {
-        this.setState({ logoutIsVisable: false });
-    };
-
-    getTittle = () => {
-        //get the current path
-        const currentPath = this.props.location.pathname;
-        let title;
-        menuList.forEach((item) => {
-            if (item.key == currentPath) {
-                title = item.title;
-            } else if (item.children) {
-                //if currentpath has the x.key starts
-                const pageNname = item.children.find(
-                    (x) => currentPath.indexOf(x.key) === 0
-                );
-                if (pageNname) title = pageNname.title;
-            }
-        });
-        return title;
-    };
-
-    render() {
-        const user = memoryUtils.user;
-        return (
-            <div className="header">
-                <Modal
-                    title="Conform"
-                    visible={this.state.logoutIsVisable}
-                    onOk={this.logout}
-                    onCancel={this.handleCancel}
+    const user = memoryUtils.user;
+    return (
+        <div className="header">
+            <Modal
+                title="Conform"
+                visible={logoutIsVisable}
+                onOk={logout}
+                onCancel={handleCancel}
+            >
+                <p>Do you want to exit?</p>
+            </Modal>
+            <div className="header-top">
+                <span>Welcome, {user.username}!</span>
+                <LinkButton
+                    onClick={() => {
+                        setLogoutIsVisable(true);
+                    }}
                 >
-                    <p>Do you want to exit?</p>
-                </Modal>
-                <div className="header-top">
-                    <span>Welcome, {user.username}!</span>
-                    <LinkButton
-                        onClick={() => {
-                            console.log('hasbeenclicked');
-                            this.setState({ logoutIsVisable: true });
-                        }}
-                    >
-                        Logout
-                    </LinkButton>
+                    Logout
+                </LinkButton>
+            </div>
+            <div className="header-bottom">
+                <div className="header-bottom-left">
+                    {headTitle}{' '}
                 </div>
-                <div className="header-bottom">
-                    <div className="header-bottom-left">
-                        {this.props.headTitle}{' '}
-                    </div>
 
-                    <div className="header-bottom-right">
-                        <span>{this.state.currentTime}</span>
-                    </div>
+                <div className="header-bottom-right">
+                    <span>{currentTime}</span>
                 </div>
             </div>
-        );
-    }
+        </div>
+    );
 }
 
-export default connect(
-    (state) => ({ headTitle: state.headTitle }),
-    {}
-)(withRouter(HeaderCom));
+const mapStateToProps = (state) => ({headTitle: state.headTitle});
+
+export default connect(mapStateToProps)(HeaderCom);
