@@ -1,12 +1,9 @@
-import React, {Component} from "react"
 import "./index.less"
 
-export default class LinkButton extends Component {
-    render() {
-        return (
-            <button className="link-button" onClick={this.props.onClick}>{
-                this.props.children
-            }</button>
-        );
-    }
+export default function LinkButton(props) {
+    return (
+        <button className="link-button" onClick={props.onClick}>{
+            props.children
+        }</button>
+    );
 }
