@@ -10,28 +10,27 @@ import {
     LineChartOutlined,
     PieChartOutlined
 } from '@ant-design/icons';
-import React from 'react';
 
 const menuList = [
     {
         title: 'Home', // menu title
         key: '/home', // path
-        icon: <HomeOutlined />, // icon name
+        icon: <HomeOutlined/>, // icon name
         isPublic: true, // public
     },
     {
         title: 'Produces',
         key: '/products',
-        icon: <AppstoreOutlined />,
+        icon: <AppstoreOutlined/>,
         children: [
             { // children menu
                 title: 'Manage category',
                 key: '/category',
-                icon: <BarsOutlined />
+                icon: <BarsOutlined/>
             }, {
                 title: 'Manage products',
                 key: '/product',
-                icon: <ToolOutlined />
+                icon: <ToolOutlined/>
             },
         ]
     },
@@ -39,29 +38,29 @@ const menuList = [
     {
         title: 'Manage users',
         key: '/user',
-        icon: <UserOutlined />
+        icon: <UserOutlined/>
     },
     {
         title: 'Manage roles',
         key: '/role',
-        icon: <SafetyCertificateOutlined />
+        icon: <SafetyCertificateOutlined/>
     }, {
         title: 'Charts',
         key: '/chart',
-        icon: <AreaChartOutlined />,
+        icon: <AreaChartOutlined/>,
         children: [
             {
                 title: 'Bar',
                 key: '/chart/bar',
-                icon: <BarChartOutlined />
+                icon: <BarChartOutlined/>
             }, {
                 title: 'Line',
                 key: '/chart/line',
-                icon: <LineChartOutlined />
+                icon: <LineChartOutlined/>
             }, {
                 title: 'Pie',
                 key: '/chart/pie',
-                icon: <PieChartOutlined />
+                icon: <PieChartOutlined/>
             },
         ]
     },
