@@ -9,15 +9,11 @@ import memoryUtils from "./utils/memoryUtils"
 
 const user = storageUtils.getUser()
 memoryUtils.user = user;
-console.log("index", user)
 
-ReactDOM.render (<React.StrictMode>
-    <Provider store={store}>
-        <App/>
-    </Provider>
-</React.StrictMode>, document.getElementById('root'));
-
-// // If you want to start measuring performance in your app, pass a function
-// // to log results (for example: reportWebVitals(console.log))
-// // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
+ReactDOM.render (
+    <React.StrictMode>
+        <Provider store={store}>
+            <App/>
+        </Provider>
+    </React.StrictMode>,
+    document.getElementById('root'));
