@@ -1,8 +1,8 @@
-import React, { Component } from 'react';
-import { Upload, Modal, message } from 'antd';
-import { PlusOutlined } from '@ant-design/icons';
-import { redeletePicture } from '../../api/action';
-import PropTypes from 'prop-types';
+import React, {Component} from 'react';
+import {Upload, Modal, message} from 'antd';
+import {PlusOutlined} from '@ant-design/icons';
+import {redeletePicture} from '../../api/action';
+
 function getBase64(file) {
     return new Promise((resolve, reject) => {
         const reader = new FileReader();
@@ -13,14 +13,10 @@ function getBase64(file) {
 }
 
 class PicturesWall extends Component {
-    static propTypes = {
-        imgs: PropTypes.array,
-    };
-
     constructor(props) {
         super(props);
         let fileList = [];
-        const { imgs } = this.props;
+        const {imgs} = this.props;
         if (imgs && imgs.length > 0) {
             fileList = imgs.map((img, index) => {
                 return {
@@ -46,7 +42,7 @@ class PicturesWall extends Component {
     }
 
     // close preview window
-    handleCancel = () => this.setState({ previewVisible: false });
+    handleCancel = () => this.setState({previewVisible: false});
     handlePreview = async (file) => {
         if (!file.url && !file.preview) {
             file.preview = await getBase64(file.originFileObj);
@@ -60,13 +56,13 @@ class PicturesWall extends Component {
         });
     };
 
-    handleChange = async ({ file, fileList }) => {
+    handleChange = async ({file, fileList}) => {
         console.log(fileList);
         if (file.status === 'done') {
             const result = file.response;
             if (result.status === 0) {
                 message.success('upload successfully');
-                const { name, url } = result.data;
+                const {name, url} = result.data;
                 console.log(result.data);
                 const file = fileList[fileList.length - 1]; // can use negative index like Python
                 file.name = name;
@@ -82,7 +78,7 @@ class PicturesWall extends Component {
                 message.error('delete failed');
             }
         }
-        this.setState({ fileList });
+        this.setState({fileList});
     };
 
     // transfor picture names to parent companent
@@ -95,12 +91,12 @@ class PicturesWall extends Component {
     };
 
     render() {
-        const { previewVisible, previewImage, fileList, previewTitle } =
+        const {previewVisible, previewImage, fileList, previewTitle} =
             this.state;
         const uploadButton = (
             <div>
-                <PlusOutlined />
-                <div style={{ marginTop: 8 }}>Upload</div>
+                <PlusOutlined/>
+                <div style={{marginTop: 8}}>Upload</div>
             </div>
         );
         return (
@@ -125,7 +121,7 @@ class PicturesWall extends Component {
                 >
                     <img
                         alt="example"
-                        style={{ width: '100%' }}
+                        style={{width: '100%'}}
                         src={previewImage}
                     />
                 </Modal>
