@@ -1,18 +1,11 @@
 import React, {Component} from 'react'
 import {Form, Input, Select} from 'antd'
-import PropTypes from 'prop-types';
 
 
 const {Item} = Form;
 const {Option} = Select;
 
 export default class AddForm extends Component {
-
-    static propTypes = {
-        showStatus: PropTypes.number.isRequired,
-        roles:PropTypes.object
-    }
-
     formRef = React.createRef();
 
     componentDidMount() {
@@ -21,25 +14,24 @@ export default class AddForm extends Component {
     }
 
     //catch the value of selector
-    onChange=(value)=>{
+    onChange = (value) => {
         this.selectorValue = value
-      }
+    }
     
-    
-    getFromValues= ()=>{
+    getFromValues = () => {
         return this.formRef.current.validateFields().then(async values => {
             // the form componment cann't capture the value in Seletor
             values.role_id = this.selectorValue
             return values;
         })
-      }
+    }
 
 
     render() {
-        const {roles,showStatus} = this.props
-        let selectedUser={}
+        const {roles, showStatus} = this.props
+        let selectedUser = {}
         // check is update or add mode
-        if(showStatus===2){
+        if (showStatus === 2) {
             selectedUser = this.props.selectedUser
         }
         return (
@@ -47,37 +39,37 @@ export default class AddForm extends Component {
                 this.formRef
             } initialValues={selectedUser}>
                 <Item name="username" label="Username" required
-                    rules={
-                        [
-                            {
-                                required: true,
-                                message: "Please input your value!"
-                            }, {
-                                min: 2,
-                                message: "must be longer than 1!"
-                            }, {
-                                max: 12,
-                                message: "Ymust be shorter than 12!"
-                            },
-                        ]
-                }>
+                      rules={
+                          [
+                              {
+                                  required: true,
+                                  message: "Please input your value!"
+                              }, {
+                              min: 2,
+                              message: "must be longer than 1!"
+                          }, {
+                              max: 12,
+                              message: "Ymust be shorter than 12!"
+                          },
+                          ]
+                      }>
                     <Input></Input>
                 </Item>
-                <Item name="password" label="Password" hidden={showStatus===2} required
-                    rules={showStatus===2?null:
-                        [
-                            {
-                                required: true,
-                                message: "Please input your value!"
-                            }, {
-                                min: 6,
-                                message: "must be longer than 1!"
-                            }, {
-                                max: 12,
-                                message: "Ymust be shorter than 12!"
-                            },
-                        ]
-                }>
+                <Item name="password" label="Password" hidden={showStatus === 2} required
+                      rules={showStatus === 2 ? null :
+                          [
+                              {
+                                  required: true,
+                                  message: "Please input your value!"
+                              }, {
+                              min: 6,
+                              message: "must be longer than 1!"
+                          }, {
+                              max: 12,
+                              message: "Ymust be shorter than 12!"
+                          },
+                          ]
+                      }>
                     <Input type='password'></Input>
                 </Item>
                 <Item name="phone" label="Phone">
@@ -88,17 +80,17 @@ export default class AddForm extends Component {
                 </Item>
                 <Item name="role_id" label="Role">
                     <Select showSearch defaultValue={selectedUser.role_id}
-                        style={
-                            {width: 200}
-                        }
-                        onChange={this.onChange}
-                        placeholder="Select a role"
-                        optionFilterProp="children"
-                        filterOption={
-                            (input, option) => option.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
-                    }>
-                        {roles.map((item,index)=>{
-                            return (<Option key = {index} value={item._id}>{item.name}</Option>)
+                            style={
+                                {width: 200}
+                            }
+                            onChange={this.onChange}
+                            placeholder="Select a role"
+                            optionFilterProp="children"
+                            filterOption={
+                                (input, option) => option.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
+                            }>
+                        {roles.map((item, index) => {
+                            return (<Option key={index} value={item._id}>{item.name}</Option>)
                         })}
                     </Select>,
                 </Item>
